+++ CHANGELOG.md
@@ -54,6 +54,8 @@
   [#4105](https://github.com/Kong/kong-operator/pull/4105)
 - Hybridgateway: add support for `konghq.com/tls-verify-depth` service annotation
   [#4099](https://github.com/Kong/kong-operator/pull/4099)
+- Hybridgateway: add support for `konghq.com/connect-timeout` service annotation
+  [#4101](https://github.com/Kong/kong-operator/pull/4101)
 - Add the following headers in requests of `ingress-controller` sent to Konnect
   for uploading configuration for tracing:
   - `X-Kic-Konnect-Sync-Instance-Id` for instance ID of Konnect config synchronizer.